+++ src/server/network/server.cpp
@@ -36,7 +36,7 @@ Server::Server(const ServerConfig& config)
 
     m_voxelData.initCommonVoxelTypes();
 
-    float seed = generateSeed("test");
+    int seed = generateSeed("test");
 
     for (int z = 0; z < m_worldSize; z++) {
         for (int x = 0; x < m_worldSize; x++) {
